Add post and comment model

+++ FirstCRUDApplication/DbEntities/Comment.cs
@@ -7,12 +7,12 @@ namespace Coffee.DbEntities
     {
         public Post Post { get; set; }
 
-        public int PostId { get; set; }
+        public long PostId { get; set; }
 
         public string Text { get; set; }
 
         public User User { get; set; }
 
-        public int UserId { get; set; }
+        public long UserId { get; set; }
     }
 }
